fixed comments and added exception handler

+++ task-service/src/main/java/com/exmple/task/service/TaskService.java
@@ -4,6 +4,8 @@ import com.exmple.task.entity.Task;
 import com.exmple.task.repository.TaskRepository;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
+import javax.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
@@ -18,38 +20,61 @@ public class TaskService {
     @Value("${tasks.find-tasks.limit}")
     private int findTasksLimit;
 
-    public long addTask(Task task) {
+    public long createTask(Task task) {
         return taskRepository.save(task).getId();
     }
 
     @Transactional
-    public boolean updateTask(Task task) {
-        boolean taskExists = taskRepository.existsById(task.getId());
-        if (taskExists) {
-            try {
-                taskRepository.updateTask(task.getId(), task.getText(), task.getTitle(), task.getTime());
-                return true;
-            } catch (Exception e) {
-                return false;
-            }
+    public void updateTaskText(final long taskId, final String text) {
+        Optional<Task> foundTask = taskRepository.findById(taskId);
+        if (foundTask.isPresent()) {
+            Task task = foundTask.get();
+            task.setText(text);
+            taskRepository.saveAndFlush(task);
+        } else {
+            throw new EntityNotFoundException("Task not found");
         }
-        return false;
     }
 
     public List<Task> getTaskByMail(String mail) {
         return taskRepository.findTaskByAuthor_Mail(mail);
     }
 
-    public boolean deleteById(long id) {
-        try {
+    public void deleteById(long id) {
+        Optional<Task> foundTask = taskRepository.findById(id);
+        if(foundTask.isPresent()) {
             taskRepository.deleteById(id);
-            return true;
-        } catch (Exception e) {
-            return false;
+        } else {
+            throw new EntityNotFoundException("Task not found");
         }
     }
 
     public List<Task> findTasksByDateAndId(final Date startDate, final long id) {
         return taskRepository.findTasksByDateAndId(startDate, id, PageRequest.of(0, findTasksLimit));
+    }
+
+    public void updateTaskTitle(final long taskId, final String title) {
+        Optional<Task> foundTask = taskRepository.findById(taskId);
+
+        if (foundTask.isPresent()) {
+            Task task = foundTask.get();
+            task.setTitle(title);
+            taskRepository.saveAndFlush(task);
+        } else {
+            throw new EntityNotFoundException("Task not found");
+        }
+    }
+
+    public void updateTaskTime(final long taskId, final Date time) {
+        Optional<Task> foundTask = taskRepository.findById(taskId);
+
+        if (foundTask.isPresent()) {
+            Task task = foundTask.get();
+            task.setTime(time);
+            taskRepository.saveAndFlush(task);
+        } else {
+            throw new EntityNotFoundException("Task not found");
+        }
+
     }
 }
